# Chunk: short-cut flatMap and traverse_ for singletons