The `traverse_` one in particular, for singleton chunks (which are generated by fs2 operations) may be useful to optimise. Chunk `flatMap` may be less relevant, but still handy.