@CremboC 2022-09-30
I added a
```scala
  @Benchmark
  def traverse(bh: Blackhole): Unit =
    bh.consume(ints.traverse(obj => Option(obj)))
```
into `ChunkBenchmark.scala` and got the following results (sorry I forgot to enable GC profiling):

before:
```
[info] Benchmark                (chunkSize)   Mode  Cnt         Score         Error  Units
[info] ChunkBenchmark.traverse            1  thrpt    5  42'933'774.352 ± 6'529'487.626  ops/s
[info] ChunkBenchmark.traverse           16  thrpt    5   3'891'490.369 ±  326'027.026  ops/s
```

after:
```
[info] Benchmark                (chunkSize)   Mode  Cnt          Score         Error  Units
[info] ChunkBenchmark.traverse            1  thrpt    5  104'065'737.195 ± 8'808'141.638  ops/s
[info] ChunkBenchmark.traverse           16  thrpt    5    3'773'273.612 ±  284'774.153  ops/s
```

I didn't ran many iterations, hence the pretty high variance, but you can see for the `traverse` change, for singleton chunks, traverse is now ~x2.5 faster.

I also ran the same test for `traverse_`:
before:
```
[info] Benchmark                                                  (chunkSize)   Mode  Cnt         Score         Error   Units
[info] ChunkBenchmark.traverse_                                             1  thrpt    5  26'243'100.321 ± 2'759'929.683   ops/s
[info] ChunkBenchmark.traverse_:·gc.alloc.rate                              1  thrpt    5      6291.698 ±     664.963  MB/sec
[info] ChunkBenchmark.traverse_:·gc.alloc.rate.norm                         1  thrpt    5       264.038 ±       0.001    B/op
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Eden_Space                     1  thrpt    5      6302.119 ±     644.001  MB/sec
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Eden_Space.norm                1  thrpt    5       264.481 ±       1.676    B/op
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Survivor_Space                 1  thrpt    5         0.045 ±       0.018  MB/sec
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Survivor_Space.norm            1  thrpt    5         0.002 ±       0.001    B/op
[info] ChunkBenchmark.traverse_:·gc.count                                   1  thrpt    5      2178.000                counts
[info] ChunkBenchmark.traverse_:·gc.time                                    1  thrpt    5      1727.000                    ms
[info] ChunkBenchmark.traverse_                                            16  thrpt    5   2'278'464.714 ±  222'078.669   ops/s
[info] ChunkBenchmark.traverse_:·gc.alloc.rate                             16  thrpt    5      6340.294 ±     620.463  MB/sec
[info] ChunkBenchmark.traverse_:·gc.alloc.rate.norm                        16  thrpt    5      3064.437 ±       0.013    B/op
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Eden_Space                    16  thrpt    5      6352.082 ±     623.893  MB/sec
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Eden_Space.norm               16  thrpt    5      3070.129 ±      11.904    B/op
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Survivor_Space                16  thrpt    5         0.051 ±       0.012  MB/sec
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Survivor_Space.norm           16  thrpt    5         0.025 ±       0.005    B/op
[info] ChunkBenchmark.traverse_:·gc.count                                  16  thrpt    5      2195.000                counts
[info] ChunkBenchmark.traverse_:·gc.time                                   16  thrpt    5      1734.000                    ms

```

after
```
[info] Benchmark                                                  (chunkSize)   Mode  Cnt          Score          Error   Units
[info] ChunkBenchmark.traverse_                                             1  thrpt    5  190'722'714.494 ± 12'942'001.137   ops/s
[info] ChunkBenchmark.traverse_:·gc.alloc.rate                              1  thrpt    5       2771.201 ±      189.585  MB/sec
[info] ChunkBenchmark.traverse_:·gc.alloc.rate.norm                         1  thrpt    5         16.002 ±        0.001    B/op
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Eden_Space                     1  thrpt    5       2774.515 ±      206.978  MB/sec
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Eden_Space.norm                1  thrpt    5         16.021 ±        0.128    B/op
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Survivor_Space                 1  thrpt    5          0.005 ±        0.009  MB/sec
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Survivor_Space.norm            1  thrpt    5         ≈ 10⁻⁵                   B/op
[info] ChunkBenchmark.traverse_:·gc.count                                   1  thrpt    5        959.000                 counts
[info] ChunkBenchmark.traverse_:·gc.time                                    1  thrpt    5        740.000                     ms
[info] ChunkBenchmark.traverse_                                            16  thrpt    5    2'304'000.620 ±   129'802.924   ops/s
[info] ChunkBenchmark.traverse_:·gc.alloc.rate                             16  thrpt    5       6410.220 ±      360.797  MB/sec
[info] ChunkBenchmark.traverse_:·gc.alloc.rate.norm                        16  thrpt    5       3064.436 ±        0.012    B/op
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Eden_Space                    16  thrpt    5       6419.960 ±      342.040  MB/sec
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Eden_Space.norm               16  thrpt    5       3069.121 ±       11.384    B/op
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Survivor_Space                16  thrpt    5          0.051 ±        0.024  MB/sec
[info] ChunkBenchmark.traverse_:·gc.churn.G1_Survivor_Space.norm           16  thrpt    5          0.024 ±        0.011    B/op
[info] ChunkBenchmark.traverse_:·gc.count                                  16  thrpt    5       2219.000                 counts
[info] ChunkBenchmark.traverse_:·gc.time                                   16  thrpt    5       1750.000                     ms
```

Which is ~7x speedup when running `traverse_` on a singleton chunk (and simultaneously half the allocations).

I think we can safely say `flatMap` is going to have performance benefits.